Create DataAdapter class
The DataAdapter class acts as a boundary between the data received from the API and the data shape the rest of the frontend codebase relies on.

By wrapping incoming API data with our DataAdapters, any changes to API data shape can be handled in one place (DataAdapters class) as opposed to having to update values everywhere.

The other benefit of using the adapter is that we can add default values for null or defined API values and transform the data for frontend use. For example, ISO date strings can be converted to formatted dates. Another example is that  undefined values such as event_name can be set as "-", which the rest of the frontend codebase can immediately render without checking the validity of the value (no more `event_name ? event_name : "-"` type checks in the react components).