Add additional functionality to MockAPIHandler
Update MockApiHandler to handle conseuctive mock API calls in a test suite (both to the same or different API endpoints).

For example, if passing in an array of resolved data `[resolvedSuccessData, resolvedErrorData]` for endpoint "/api/users":
- the first call to "/api/users" will return resolvedSuccessData
- the second call and all subsequent calls to "/api/users" will return resolvedErrorData.
MockApiHandler will remove the first item of the resolved data array after each call, except if the there is only one item in the array (in which case, the data will persist and be available for any subsequent calls).

- Update creation of new MockApiHandler instance to expect an object of `{ [endpoint]: [array of resolved data, whether error or success] }` shape.
- Create Endpoints class to consolidate all formation of endpoint strings
- Create MockApiData class to readily create correct success and error responses

Example MockApiHandler usage
```
        const successfulLoginData = MockApiData.successfulLoginData({ uuid: uuid })
       const failedLoginData = MockApiData.errorData(["Issue logging in"])
        const clientsList = MockApiData.successData([]);
        const apiHandler = new MockAPIHandler({
            [Endpoints.getUser(uuid)]: [successfulLoginData, failedLoginData],
            [Endpoints.getClients(uuid)]: [clientsList]
        });
```