Create EventInformation view
If there is event information, it will render the wedding or default event modules depending on event name (it will look for "wedding" in the event_name to trigger wedding specific fields)

If there is no event information tied to a client, it will render empty event modules when a user  updates the Package information, specifically the "engagement_included" and "wedding_included" fields. It will render an engagement or wedding event module depending on what the package wants to include.  This is to prepare for editing of these values in the UpdateEventForm module to be created in the next story

- Update Event DataAdapter with new event fields. It can also return a set of initial empty values to prepare for event creation.