Refactor & Fix Form Values Bug
Issue:
The client related mdoules are currently endering a "-" in the place of undefined or null values. While this is great, its a poor experience to pass those placeholder values into the Form modules. For example, a user would then have to delete the "-" before typing in a new value.

In the efforts of transforming the data to be form ready, there was a strange bug in which the form would have no value as expected, but closing the form and/or saving the form would cause the VIew modules on the client page to render with no placeholder values as well (no "-").  It seemed as if the transformed values from the API was also getting effected (rendering with no default value), until a page refresh hits and then it would render as expected again.

Fix:
This is all actually related to objects still referenced to each other. A data object passed into form and the related view components were references to an original. So trying to adjust values to what I thought was a clean copy of the values were actually effecting all copies being used everywhere.

Once I realized this, the fix is to make sure any transformation functions work off a COPY of a passed in value. Essentially using something like `Object.assign()` to create a new copy, and then make sure not to change anything in the original passed in value.

This needed to go in effect in the new DataAdapter.toFormReadyData() and DataAdapter.toApiReadyClient() function. Surprisingly, this was effecting the ApiReadyClient function in hidden ways but it was not an obvious bug. For example, after an API call, we receive console warnings saying the input fields were an ISOstring as opposed to the expected data type. This was because the transformation was effecting the original data.

This commit also includes a Large folder restructuring:
- Extracts components from Client into their own folders.
- Move UI under src
- Rename APIHandler folder to "api" and move related api utilities into it