Update Event Creation Flow
Previously:
An empty event module is created when a Package is set to include weddings or engagements. For example, toggling that an engagement is included in the package would render a empty uncreated Engagement event module. However, there were 2 parts to this flow that led to a bug:
(1) the empty uncreated event modules look exactly the same as a created event module, which made users think it was real
(2) the empty event module, lets say for an Engagement event,  would allow the user to change the event name from Engagement to Wedding. When the event was saved, it would confuse the users because they would then see another Engagement module along with the newly created wedding one.
The system was trying to be helpful by letting the user realize they haven't created a new intended Engagement event. But this lead to lots of confusion and it then became out of sync with the backend API. Specifically, the user created a wedding event without toggling the wedding inclusion in Package. Any new updates to the package does not delete the new Wedding event even though it should be deleted as its not included in the package!

Fix:
The fix is to revise the UX flow to help prevent the numerous bugs that came from the previous event creation flow.
Now for example, toggling engagement inclusion in the package will render a BUTTON with "Create Engagement Event" as opposed to an empty event module.
When the user clicks the button, the form will only allow users to create the event with the exact event name specified, in this case "Engagement".

- Creation of NewSuggestedEvent and ExistingEvent modules.
- Update responsive design of Client page
- Update client labels to be a dropdown of options.